Revert "added mapping function in commonHelper"
This reverts commit 7971d3df0fa385a5b8e29e75f763aa9749126c36.
