+++ integration_tests/tests/e2e-tests.ts
@@ -1,7 +1,6 @@
 import {
   generateTestEvent,
   getYearMonth,
-  mapBillingTransactionCurated,
   poll,
   TableNames,
 } from "../../src/handlers/int-test-support/helpers/commonHelpers";
@@ -19,6 +18,7 @@ import {
   TestData,
   TestDataRetrievedFromConfig,
 } from "../../src/handlers/int-test-support/helpers/testDataHelper";
+import { BillingTransactionCurated } from "./billing-and-transaction-view-tests";
 import crypto from "crypto";
 
 const prefix = resourcePrefix();
@@ -144,7 +144,18 @@ export const assertQueryResultWithTestData = async (
     eventTime
   );
 
-  const response = mapBillingTransactionCurated(curatedResponse);
+  const response: BillingTransactionCurated = curatedResponse.map((item) => {
+    return {
+      vendor_id: item.vendor_id,
+      vendor_name: item.vendor_id,
+      service_name: item.service_name,
+      year: item.year,
+      month: item.month,
+      billing_price_formatted: item.billing_price_formatted,
+      transaction_price_formatted: item.transaction_price_formatted,
+      price_difference_percentage: item.price_difference_percentage,
+    };
+  });
   expect(response.length).toBe(1);
   expect(response[0].billing_price_formatted).toEqual(
     expectedResults.billingPriceFormatted
